Move backend/utils/pubsub.js one directory up
